Create Model for efficacy against symptomatic COVID-19
This is the Bayesian Network for the symptomatic model. We have defined an "assault" probability (pA), that takes into account all possible factors that determine the probability of becoming infected. This probability is the same for the vaccinated and placebo and creates an additional layer in the graph where we have the unobserved amounts of the “assaulted” number of infected in the two groups of vaccinated (nVa ) and placebo (nPa ). Assaulted and infected nodes are linked to their parents via a binomial distribution and we have chosen a flat prior for pA and the efficacy of the vaccine.