Create Efficacy against severe disease
In this other model, the inference about efficacy depends on our knowledge of the number of severals in the placebo case. Astrazeneca has not released data for the number of severe placebos and therefore it is interesting to wonder how the confidence in the distribution of vaccine efficacy changes as this number varies. We modeled the non-knowledge of the data through our a priori knowledge of the phenomenon. In particular, we assumed that the number of severe placebos was a Gaussian around a certain mean value, and we have studied the variation of the efficacy respect to the variation of this mean value. If the number of severe placebos becomes small enough, the modal value of the distribution of the efficacy is no longer 100% and our confidence in the claimed results gets smaller. For this reason it would have been appropriate for Astrazeneca to publish that data.